+++ src/optimization/estimation/newton/Newton.cpp
@@ -6,35 +6,35 @@
 
 
 // The Newton functions are hard-coded
-std::vector<float> Newton::gradient(const std::vector<float> &position) {
-    std::vector<float> grad = {2 * (position[0] + 3), 1.0f};
+std::vector<double> Newton::gradient(const std::vector<double> &position) {
+    std::vector<double> grad = {2 * (position[0] + 3), 1.0};
     return grad;
 }
 
-std::vector<float> Newton::hessian(const std::vector<float> &position) {
-    std::vector<float> hess = {2.0f, 0.0f, 0.0f, 0.0f};
+std::vector<double> Newton::hessian(const std::vector<double> &position) {
+    std::vector<double> hess = {2.0, 0.0, 0.0, 0.0};
     return hess;
 }
 
 // The determinant is used to calculate the inverse of the Hessian
-float Newton::determinant(const std::vector<float> &hess) {
-    float det = hess[0] * hess[3] - hess[1] * hess[2];
+double Newton::determinant(const std::vector<double> &hess) {
+    double det = hess[0] * hess[3] - hess[1] * hess[2];
     return det;
 }
 
-std::vector<float> Newton::step(const std::vector<float> &grad, const std::vector<float> &hess, const float det, const Particle &particle) {
+std::vector<double> Newton::step(const std::vector<double> &grad, const std::vector<double> &hess, const double det, const Particle &particle) {
     // step = inverse of the Hessian * gradient
-    float step_x = (hess[3] * grad[0] - hess[1] * grad[1]) * ( 1 / det);
-    float step_y = (- hess[2] * grad[0] + hess[0] * grad[1]) * ( 1 / det);
+    double step_x = (hess[3] * grad[0] - hess[1] * grad[1]) * ( 1 / det);
+    double step_y = (- hess[2] * grad[0] + hess[0] * grad[1]) * ( 1 / det);
 
-    // float step_x = -1.0 / hess[0] * grad[0] - 1.0 / hess[1] * grad[1];
-    // float step_y = -1.0 / hess[2] * grad[0] -  1.0 / hess[3] * grad[1];
+    // double step_x = -1.0 / hess[0] * grad[0] - 1.0 / hess[1] * grad[1];
+    // double step_y = -1.0 / hess[2] * grad[0] -  1.0 / hess[3] * grad[1];
 
-    std::vector<float> new_position = {particle.get_position()[0] - step_x, particle.get_position()[1] - step_y};
+    std::vector<double> new_position = {particle.get_position()[0] - step_x, particle.get_position()[1] - step_y};
     return new_position;
 }
 
-Newton::Newton(const int n_particles, const int rank, const int max_iter, const float threshold,
+Newton::Newton(const int n_particles, const int rank, const int max_iter, const double threshold,
                std::vector<float> &min_pos, std::vector<float> &max_pos) : Optimization(n_particles, rank, max_iter,
                                                                                         threshold, min_pos, max_pos) {
     initialize_optimization();
@@ -42,10 +42,10 @@ Newton::Newton(const int n_particles, const int rank, const int max_iter, const 
 }
 
 void Newton::run() {
-    std::vector<float> grad;
-    std::vector<float> hess;
-    float det;
-    std::vector<float> new_position;
+    std::vector<double> grad;
+    std::vector<double> hess;
+    double det;
+    std::vector<double> new_position;
 
     int stopping_counter = 0;
     float stop_condition = global_best.get_fitness();
